Move `gossip` module under `source`
